requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
dictdiffer==0.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Tortoise==0.0.2
Flask-UUID==0.2
importlib_metadata==8.5.0
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nc-console==0.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
Werkzeug==3.0.4
zipp==3.20.2
